Fix unarchive-or-bin sql command in inbox
The command as it was currently implemented would not move a
conversation to the bin when aggregating it onto the set_inbox commands,
as the sql query was so that it would only unarchive entries on request,
but the entry was not on the archive, so the box was not moved.

This happens because we enforce automatic moves from the archive to the
inbox, and no other box is to be moved automatically.